+++ modules/agar/CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-- Added new keys pageUp and pageDown.
+- Added `pageUp` and `pageDown` constants to the `Keys` API
 
 ## 4.13.1 - 2019-05-20